Add support for smart selection of statements.
This change modifies Stencil and Transformer selection of statement source text.
Instead of just extracting the text corresponding to the statement source range,
they now use a context-sensitve analysis to determine the range.  Currently,
this only applies to statements: those with an associated trailing ';' are now
extracted with their ';'.  However, we intend to extend this in the future to
more nodes, to generally select the "intuitive" range, rather than the one
strictly defined by the node.